Process IDs (lldb::pid_t) and thread IDs (lldb::tid_t) are now 64 bit. This will allow us to represent a process/thread ID using a pointer for the OS plug-ins where they might want to represent the process or thread ID using the address of the process or thread structure.
llvm-svn: 145644

+++ lldb/include/lldb/lldb-types.h
@@ -97,8 +97,8 @@ namespace lldb
 {
     typedef uint64_t    addr_t;
     typedef uint64_t    user_id_t;
-    typedef int32_t     pid_t;
-    typedef uint32_t    tid_t;
+    typedef uint64_t    pid_t;
+    typedef uint64_t    tid_t;
     typedef int32_t     break_id_t;
     typedef int32_t     watch_id_t;
     typedef void *      clang_type_t;
